// cypress/support/pageObjects/dataEntryForm/interfaces/datePickersInterface.ts
export interface DatePickersLocatorsInterface {
  datePickerOneCurrentDayLocator: string;
  datePickerOneLocator: string;
  datePickerTwoLocator: string;
  datePickerThreeStartDateInputLocator: string;
  datePickerThreeEndDateInputLocator: string;
  datePickerThreeSubmitButton: string;
  resultLocator: string;
  resultTextLocator: string;
};  